+++ tests/tests_unit/rules/test_importers/test_dtdl_importer.py
@@ -5,7 +5,7 @@ from cognite.neat._issues.errors import ResourceMissingIdentifierError, Resource
 from cognite.neat._issues.warnings import PropertyTypeNotSupportedWarning, ResourceTypeNotSupportedWarning
 from cognite.neat._rules.importers import DTDLImporter
 from cognite.neat._rules.importers._dtdl2rules.spec import DTMI, Interface
-from cognite.neat._rules.models import InformationRules, SchemaCompleteness
+from cognite.neat._rules.models import InformationRules
 from cognite.neat._rules.transformers import ImporterPipeline
 from tests.tests_unit.rules.test_importers.constants import DTDL_IMPORTER_DATA
 
@@ -56,7 +56,6 @@ class TestDTDLImporter:
                     }
                 )
             ],
-            schema=SchemaCompleteness.complete,
         )
 
         result = ImporterPipeline.try_verify(dtdl_importer)