+++ tests/tests_unit/rules/test_importers/test_imf_importer.py
@@ -19,12 +19,3 @@ def test_imf_importer():
         class_property_pairs[ClassEntity.load("pcaimf:IMF_1ccc23fc_42ca_4b8a_acd5_ef2beddf7f12")]["hasTerminal"].type_
         == EntityTypes.object_property
     )
-
-    assert (
-        str(
-            class_property_pairs[ClassEntity.load("pcaimf:IMF_1ccc23fc_42ca_4b8a_acd5_ef2beddf7f12")][
-                "hasTerminal"
-            ].transformation
-        )
-        == "prefix-3:1ccc23fc-42ca-4b8a-acd5-ef2beddf7f12(prefix-6:hasTerminal)"
-    )