+++ tests/tests_unit/rules/test_models/test_information_rules.py
@@ -75,7 +75,6 @@ def invalid_domain_rules_cases():
         {
             "Metadata": {
                 "role": "information architect",
-                "schema": "complete",
                 "creator": "Jon, Emma, David",
                 "space": "power",
                 "external_id": "power2consumer",
@@ -88,9 +87,7 @@ def invalid_domain_rules_cases():
                 {
                     "Class": "GeneratingUnit",
                     "Description": None,
-                    "Parent Class": None,
-                    "Source": "http://www.iec.ch/TC57/CIM#GeneratingUnit",
-                    "Match": "exact",
+                    "Implements": None,
                 }
             ],
             "Properties": [
@@ -102,8 +99,6 @@ def invalid_domain_rules_cases():
                     "Min Count": 1,
                     "Max Count": 1.0,
                     "Default": None,
-                    "Source": None,
-                    "MatchType": None,
                     "Transformation": ":GeneratingUnit(cim:name)",
                 }
             ],
@@ -131,9 +126,7 @@ def incomplete_rules_case():
                 {
                     "Class": "GeneratingUnit",
                     "Description": None,
-                    "Parent Class": None,
-                    "Source": "http://www.iec.ch/TC57/CIM#GeneratingUnit",
-                    "Match": "exact",
+                    "Implements": None,
                 }
             ],
             "Properties": [
@@ -145,18 +138,22 @@ def incomplete_rules_case():
                     "Min Count": 1,
                     "Max Count": 1.0,
                     "Default": None,
-                    "Source": None,
-                    "MatchType": None,
-                    "Rule Type": "rdfpath",
-                    "Rule": "cim:GeneratingUnit",
+                    "Transformation": "cim:GeneratingUnit",
                 }
             ],
         },
-        ResourceNotDefinedError[ClassEntity](
-            ClassEntity(prefix="power", suffix="GeneratingUnit2"),
-            "class",
-            "Classes sheet",
-        ),
+        [
+            ResourceNotDefinedError[ClassEntity](
+                ClassEntity(prefix="power", suffix="GeneratingUnit2"),
+                "class",
+                "Classes sheet",
+            ),
+            ResourceNotDefinedError[ClassEntity](
+                ClassEntity(prefix="power", suffix="GeneratingUnit"),
+                "class",
+                "Classes sheet",
+            ),
+        ],
         id="missing_rule",
     )
 
@@ -188,7 +185,8 @@ class TestInformationRules:
         with pytest.raises(ValueError) as e:
             InformationRules.model_validate(incomplete_rules)
         errors = NeatError.from_pydantic_errors(e.value.errors())
-        assert errors[0] == expected_exception
+        assert len(errors) == 2
+        assert set(errors) == set(expected_exception)
 
     @pytest.mark.parametrize("rules, expected_exception", list(case_insensitive_value_types()))
     def test_case_insensitivity(self, rules: dict[str, dict[str, Any]], expected_exception: DataType) -> None:
@@ -200,7 +198,7 @@ class TestInformationRules:
 
         assert isinstance(dms_rules, DMSRules)
 
-    @pytest.mark.skip(reason="Skipping this test until rules refactoring is completed")
+    @pytest.mark.skip("Not sure purpose of this test, so skipping for now")
     def test_olav_as_dms(self, olav_rules: InformationRules) -> None:
         olav_rules_copy = olav_rules.model_copy(deep=True)
         # Todo: Remove this line when Olav's Information .xlsx file is available
@@ -209,9 +207,10 @@ class TestInformationRules:
             if cls_.class_.versioned_id == "power_analytics:GeoLocation":
                 continue
             elif cls_.class_.versioned_id in ("power_analytics:Point", "power_analytics:Polygon"):
-                cls_.parent = None
+                cls_.implements = None
             new_classes.append(cls_)
         olav_rules_copy.classes = new_classes
+
         ## End of temporary code
         dms_rules = InformationToDMS().transform(olav_rules_copy).rules
 
@@ -263,42 +262,9 @@ class TestInformationRulesConverter:
 
         assert actual_space == expected_space
 
-    @pytest.mark.skip(
-        reason="Skipping this test until rules refactoring is completed, probably deprecating as we do not use last"
-    )
-    def test_svein_harald_information_as_dms(self, svein_harald_information_rules: InformationRules) -> None:
-        expected = {
-            "ArrayCable": {"PowerLine"},
-            "DistributionLine": {"PowerLine"},
-            "DistributionSubstation": {"Substation"},
-            "ElectricCarCharger": {"EnergyConsumer"},
-            "ExportCable": {"PowerLine"},
-            "MultiLineString": {"GeoLocation"},
-            "OffshoreSubstation": {"Substation"},
-            "OnshoreSubstation": {"TransmissionSubstation"},
-            "Point": {"GeoLocation"},
-            "Polygon": {"GeoLocation"},
-            "Transmission": {"PowerLine"},
-            "TransmissionSubstation": {"Substation"},
-            "WindFarm": {"EnergyArea"},
-            "WindTurbine": {"GeneratingUnit"},
-        }
-        dms_rules = InformationToDMS().transform(svein_harald_information_rules).rules
-
-        assert isinstance(dms_rules, DMSRules)
-        assert dms_rules.last is not None
-        actual = {
-            view.view.external_id: {parent.external_id for parent in view.implements}
-            for view in dms_rules.last.views
-            if view.implements
-        }
-
-        assert actual == expected
-
     def test_convert_above_container_limit(self) -> None:
         info = InformationInputRules(
             metadata=InformationInputMetadata(
-                schema_="complete",
                 space="bad_model",
                 external_id="bad_model",
                 name="Bad Model",